Replaced supposedly broken JsonFormatter with JsonSparseFormatter
