+++ src/reader/FormatProcessor.cpp
@@ -34,7 +34,6 @@
 
 #include "caliper/reader/Expand.h"
 #include "caliper/reader/JsonFormatter.h"
-#include "caliper/reader/JsonSparseFormatter.h"
 #include "caliper/reader/TableFormatter.h"
 #include "caliper/reader/TreeFormatter.h"
 #include "caliper/reader/UserFormatter.h"
@@ -55,17 +54,15 @@ const char* tree_kernel_args[]   = { "path-attributes" };
 enum FormatterID {
     Csv         = 0,
     Json        = 1,
-    JsonSparse  = 2,
-    Expand      = 3,
-    Format      = 4,
-    Table       = 5,
-    Tree        = 6
+    Expand      = 2,
+    Format      = 3,
+    Table       = 4,
+    Tree        = 5
 };
 
 const QuerySpec::FunctionSignature formatters[] = {
     { FormatterID::Csv,    "csv",    0, 0, nullptr },
     { FormatterID::Json,   "json",   0, 0, nullptr },
-    { FormatterID::JsonSparse,   "json-sparse",   0, 0, nullptr },
     { FormatterID::Expand, "expand", 0, 0, nullptr },
     { FormatterID::Format, "format", 1, 2, format_kernel_args },
     { FormatterID::Table,  "table",  0, 0, nullptr },
@@ -105,10 +102,7 @@ struct FormatProcessor::FormatProcessorImpl
                 m_formatter = new CsvFormatter(m_stream);
                 break;
             case FormatterID::Json:
-                m_formatter = new JsonFormatter(spec);
-                break;
-            case FormatterID::JsonSparse:
-                m_formatter = new JsonSparseFormatter(m_stream, spec);
+                m_formatter = new JsonFormatter(m_stream, spec);
                 break;
             case FormatterID::Expand:
                 m_formatter = new Expand(m_stream, spec);
